Implent Material Design and a few other things (v2)
Ihave implemented a few changes and they are as follows.

1. Implenented
a color pallete provided
by https://www.materialpalette.com/blue/light-blue. You can now try out
new app themes by just creating a new one at the www.materialpallete.com
site and downloading the XML and then just copy/pasting into the
colors.xml file.

2. MainActivity Fab button color changed to primary
due to contrast conflicts with the black boarder around the current
icon. You can try
changing it to primary_light to see what I mean. If this is your
preference then it is ok.

3. Changed the background color of the Fab buttons inthe ArcMenu to
primary_light. These could just as well be primary instead of
primary_light also.

4. Added clipChildren="false" to the ArcMenu so that the shadows of the
children fAB buttons aren't clipped in the Gear  screen. This did not
completely solve the problem so I added a marginLeft to the "suit" fAB
and a marginTop to the "rig" fAB for symmetry. There is additional
clipping during the animations which can be handled in another commit.

5. I changed the colors in the gradient in the side_nav_bar drawable to
bring them in line with the overall app theme.

6. Changed the color of the progress bar to bring it more in line with
the overall app theme.

7. Changed a few bottom|right values to bottom|end for compatibility.

There are a number of xml tags that are designed for API 21+ while your
minimum supported API is currently set at 19 and could cause
compatibility errors. Could also be handled in another Issue.
I left the colors in tact that are used when setting up the JumpRecycler
and the JumpView (black and grey) becuase I am not sure if you intended
for app theme color changes to affect colors that are used when setting
up pre-populated form data.

One thing I noticed is that the text values in the res/layout files are
not stored in string.xml which may cause issues with translation. This
could be done in a separate issue.

If you need anything else further done or anything changed feel free to
let me know. It's no problem. :)