Fixed the number of Exits being displayed
The issue was with the count() being called in SQLiteModel. We must use
the other method that takes in a HashMap<> to perform the proper query
operation on the Sql database. Adding Synchronizable.getActiveParams()
accomplishes this task by returning only the database rows where deleted
is not 0.

P.S. Also fixed an indentation in fragment_jump_view! :+1: